+++ source/extensions/lightspeed.trex.stage_manager.plugin.interaction.usd/lightspeed/trex/stage_manager/plugin/interaction/usd/all_meshes.py
@@ -41,6 +41,7 @@ class AllMeshesInteractionPlugin(_StageManagerUSDInteractionPlugin):
     ]
     # TODO StageManager: We have LSS plugin names in the flux ext because of this system
     compatible_widgets: list[str] = [
+        "AssignCategoryActionWidgetPlugin",
         "CustomTagsWidgetPlugin",
         "FocusInViewportActionWidgetPlugin",
         "IsCaptureStateWidgetPlugin",