+++ source/extensions/lightspeed.trex.stage_manager.plugin.widget.usd/lightspeed/trex/stage_manager/plugin/widget/usd/extension.py
@@ -20,6 +20,7 @@ import carb.settings
 import omni.ext
 from omni.flux.stage_manager.factory import get_instance as _get_factory_instance
 
+from .action_assign_category import AssignCategoryActionWidgetPlugin as _AssignCategoryActionWidgetPlugin
 from .action_remap_skeleton import RemapSkeletonActionWidgetPlugin as _RemapSkeletonActionWidgetPlugin
 from .focus_in_viewport import FocusInViewportActionWidgetPlugin as _FocusInViewportActionWidgetPlugin
 from .info_remap_skeleton import RemapSkeletonInfoWidgetPlugin as _RemapSkeletonInfoWidgetPlugin
@@ -35,6 +36,7 @@ class LightspeedStageManagerUSDWidgetPluginsExtension(omni.ext.IExt):
         _RemapSkeletonActionWidgetPlugin,
         _RemapSkeletonInfoWidgetPlugin,
         _IsCategoryHiddenStateWidgetPlugin,
+        _AssignCategoryActionWidgetPlugin,
     ]
 
     def on_startup(self, _):