+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_skeletons.py
@@ -40,6 +40,7 @@ class AllSkeletonsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     ]
     # TODO StageManager: We have LSS plugin names in the flux ext because of this system
     compatible_widgets: list[str] = [
+        "AssignCategoryActionWidgetPlugin",
         "CustomTagsWidgetPlugin",
         "FocusInViewportActionWidgetPlugin",
         "IsCaptureStateWidgetPlugin",