+++ source/extensions/lightspeed.trex.mesh_properties.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.2]
+## Changed
+- Moved the categories dialog
+
 ## [1.2.1]
 ## Changed
 - Set the column widths and allow resizing