+++ source/extensions/lightspeed.trex.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.4.5]
+## Added
+- Added an action for categories
+
 ## [1.4.4]
 ## Changed
 - Update to Kit 106.5