+++ source/extensions/lightspeed.trex.utils.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.1.0]
+### Added
+- Moved the categories dialog
+
 ## [1.0.3]
 ### Changed
 - Changed repo link